pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codewash"
version = "0.1.0"
description = "Anonymize code repositories for safe AI processing"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "codewash contributors", email = "opensource@example.com" }
]
keywords = [
    "anonymization",
    "security",
    "infrastructure",
    "devops",
    "ai",
    "credentials",
    "secrets",
    "privacy",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Security",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/mxfnk/codewash"
Repository = "https://github.com/mxfnk/codewash.git"
Documentation = "https://github.com/mxfnk/codewash#readme"
Issues = "https://github.com/mxfnk/codewash/issues"
Changelog = "https://github.com/mxfnk/codewash/blob/main/CHANGELOG.md"

[project.scripts]
codewash = "codewash.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/codewash"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
exclude = ["tests/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
